refactoring directory consituetion
