+++ data_edit/graph/module.py
@@ -2,11 +2,7 @@ import glob, os
 import pandas as pd
 import matplotlib.pyplot as plt
 from file_operation.module import *
-from statistics import mean
-from tabulate import tabulate
-
-#dataframeがあればよいので、csvをひたすらに読み込んで(usecolsは瞬目の間隔時間平均だけ)瞬目の間隔時間平均のデータを取得する。
-#読み込む際は、経過時間に対応している疲労度も一緒に読み込み、どの疲労度の時の瞬目の間隔時間平均か求めていく
+from data_edit.module import get_date, get_fatigue_data_path, get_fatigue_data
 
 def create_graph_from_ditection_data(alt_date, hours, result, colums, threshold):
     #グラフの初期設定
@@ -130,7 +126,6 @@ def create_graph_from_ditection_data_ready(alt_date, hours, colums, jins_meme_da
                 #疲労度に関係のある指標を取得(基本はstrongBlinkIntervalAvgを使っている)
                 fatigue_relation_value = get_fatigue_relation_value(jins_meme_data_name, per_minitue_data)
 
-                #TODO:この部分もget_fatigue_data_pathで取得するようにするか。minutueの定義もまとめられるし
                 fatigue_data_file_path = get_fatigue_data_path(per_minitue_data[0][0])
                 #5分ごとに疲労度のデータ取得
                 if pass_time % 5 == 0 and os.path.exists(fatigue_data_file_path):
@@ -138,18 +133,12 @@ def create_graph_from_ditection_data_ready(alt_date, hours, colums, jins_meme_da
 
                 #1分ごとの経過時間・瞬目の間隔時間平均・疲労度のレコードを作成
                 date_blink_interval_time_fatigue_data.append([pass_time, fatigue_relation_value, fatigue])
+                
                 #瞬目の間隔時間平均の閾値を定める
                 if fatigue >= 3 and not threshold:
                     threshold  = {'pass_time': pass_time, 'fatigue_relation_value': fatigue_relation_value}
                 
     return date_blink_interval_time_fatigue_data, threshold 
-
-""" 日付から年月日それぞれに分ける """
-def get_date(date):
-    year = date[0:4]
-    month = date[4:6]
-    day = date[6:8]
-    return year, month, day
 
 """ csvファイルにある疲労度に関係のある指標を取得する """
 def get_fatigue_relation_value(jins_meme_data_name, per_minute_data):
@@ -165,12 +154,6 @@ def get_fatigue_relation_value(jins_meme_data_name, per_minute_data):
         value = per_minute_data[0][1]
     
     return value
-
-""" 疲労度の値を取得 """
-def get_fatigue_data(file_path):
-    with open(file_path, 'r') as file:
-        value = file.read()
-    return int(value)
 
 """ グラフ作成時に「休憩あり」か「休憩なし」か入力する際のバリデーション """
 def check_rest_or_non(input):
@@ -206,87 +189,3 @@ def get_shaft_interval_figsize(less_than_one_hour_flag):
         figsize = (12, 9)
     
     return per_five_minute, per_ten_minute, figsize
-
-
-
-""" 疲労度ごとの瞬目の間隔時間平均の振り幅の表作成 """
-def create_blink_interval_time_amplitude_table(date, jins_meme_data_name):
-    #疲労度ごとの瞬目の間隔時間平均値のhashを取得
-    blink_interval_at_fatigue_hash = ready_blink_interval_at_fatigue_hash(date, jins_meme_data_name)
-    #取得したhashの分析
-    blink_interval_at_fatigue_minimum_max, blink_interval_at_fatigue_average = analyze_blink_interval_at_fatigue(blink_interval_at_fatigue_hash)
-    
-    table_data_average = get_table_data(blink_interval_at_fatigue_average, default_flag=True)
-    table_data_minimum_max = get_table_data(blink_interval_at_fatigue_minimum_max)
-    headers = ['1', '2', '3', '4', '5']
-
-    table_minimum_max = tabulate(tabular_data=table_data_minimum_max, headers=headers, tablefmt='html')
-    table_average = tabulate(tabular_data=table_data_average, headers=headers, tablefmt='html')
-    
-    
-    if input("疲労度ごとの瞬目の間隔時間平均値の分析結果を表示しますか？(yes or no)\n") == 'yes':
-        print('[最小値, 最大値]')
-        print(blink_interval_at_fatigue_minimum_max)
-        print('平均値')
-        print(blink_interval_at_fatigue_average)
-        
-    return table_minimum_max, table_average
-
-""" 疲労度のパスを取得 """
-def get_fatigue_data_path(date_time):
-    year = date_time[0:4]
-    month = date_time[5:7]
-    day = date_time[8:10]
-    hour_in_japan = str(int(date_time[11:13]) + 9).zfill(2)
-    minute = date_time[14:16]    
-    fatigue_data_path = f'./fatigue_data/{year}/{month}/{day}/{hour_in_japan}/{minute}.txt'
-    
-    return fatigue_data_path
-
-""" 疲労度ごとの瞬目の間隔時間平均の関係を作成(dict形式) """
-def ready_blink_interval_at_fatigue_hash(date, jins_meme_data_name):
-    year, month, _ = get_date(date)
-    target_pathes = glob.glob(f'./ditection_data/{year}/{month}/*/*/*{jins_meme_data_name}.csv')
-    colums = ['date', 'strongBlinkIntervalAvg']
-    date_blink_interval_list = readCsv(target_pathes, colums)
-    
-    #瞬目の間隔時間平均と疲労度の関連付けの初期化
-    blink_interval_at_fatigue_hash = {}
-    for key in range(1, 6):
-        blink_interval_at_fatigue_hash[key] = []
-
-    for date_blink_interval in [value[0] for value in date_blink_interval_list]:
-        if len(date_blink_interval) != 0:
-            fatigue_data_path = get_fatigue_data_path(date_blink_interval[0])
-            if os.path.exists(fatigue_data_path):
-                fatigue = get_fatigue_data(fatigue_data_path)
-                #特殊なフィルター(不正なデータを取り除く)
-                if fatigue == 2 and date_blink_interval[1] > 7:
-                    continue
-                blink_interval_at_fatigue_hash[fatigue].append(date_blink_interval[1])
-    
-    return blink_interval_at_fatigue_hash
-
-""" 疲労度ごとの瞬目の間隔時間平均値を分析 """
-def analyze_blink_interval_at_fatigue(blink_interval_at_fatigue_hash):
-    #最大値最小値を求める
-    for key, values in blink_interval_at_fatigue_hash.items():
-        blink_interval_at_fatigue_hash[key] = [min(values), max(values)]
-    blink_interval_at_fatigue_max_minimum = blink_interval_at_fatigue_hash.copy()
-    for key, values in blink_interval_at_fatigue_hash.items():
-        blink_interval_at_fatigue_hash[key] = mean(values)
-    blink_interval_at_fatigue_average = blink_interval_at_fatigue_hash.copy()
-
-    return blink_interval_at_fatigue_max_minimum, blink_interval_at_fatigue_average
-
-""" 表作成するために加工したデータを取得 """
-def get_table_data(hash, default_flag=False):
-    if default_flag:
-        table_data = [list(hash.values())] 
-    else:
-        array = []
-        for values in hash.values():
-            array.append(' ~ '.join([str(value) for value in values]))
-        table_data = [array]
-    
-    return table_data
